+++ src/js/@llrt/runtime.ts
@@ -198,7 +198,6 @@ const startProcessEvents = async (
       }
       iterations++;
     }
-    global.__gc();
   }
 };