+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Use MiMalloc allocator for improved performance.
+* Use SnMalloc allocator for improved performance.
 * Switch libc from gnu to musl
 * Bug fixes
 * Dependency upgrades